Cleaned recipient_account.py of static methods and moved it to types package, adjusted recipient_account_gateway imports accordingly and added more inline comments.
