+++ test/integration/RecipientTest.py
@@ -7,7 +7,7 @@ sys.path.append(os.path.abspath('.'))
 
 from trolley.configuration import Configuration
 from trolley.recipient import Recipient
-from trolley.recipient_account import RecipientAccount
+from trolley.types.recipient_account import RecipientAccount
 from TestSetup import TestSetup
 from TestHelper import TestHelper
 from trolley.exceptions.malformedException import MalformedException
