+++ trolley/recipient_account_gateway.py
@@ -1,6 +1,7 @@
 from collections import namedtuple
 from trolley.exceptions.invalidFieldException import InvalidFieldException
 import trolley.configuration
+from trolley.types.recipient_account import RecipientAccount
 
 
 class RecipientAccountGateway(object):
@@ -13,6 +14,11 @@ class RecipientAccountGateway(object):
         self.gateway = gateway
         self.config = config
 
+    """
+        Retrieve all the recipient accounts
+            A recipient_id is required::
+            recipient_account.findAll('R-fjeracjmuflh')
+        """
     def findAll(self, recipient_id):
         if recipient_id is None:
             raise InvalidFieldException("Recipient id cannot be None.")
@@ -23,7 +29,7 @@ class RecipientAccountGateway(object):
         recipientaccounts = []
         count = 0
         for recipientaccount in response['accounts']:
-            recipaccount = trolley.recipient_account.RecipientAccount.factory(
+            recipaccount = RecipientAccount.factory(
                 recipientaccount)
             newrecipientaccount = namedtuple(
                 "RecipientAccount", recipaccount.keys())(*recipaccount.values())
@@ -31,42 +37,63 @@ class RecipientAccountGateway(object):
             count = count + 1
         return recipientaccounts
 
+    """
+        Retrieve a recipient account
+            A recipient_id and recipient_account_id are required::
+            recipient_account.find('R-fjeracjmuflh','A-2DQMpN4jurTFn9gRxobx4C')
+        """
     def find(self, recipient_id, recipient_account_id):
         if recipient_id is None:
             raise InvalidFieldException("Recipient id cannot be None.")
         endpoint = '/v1/recipients/' + recipient_id + '/accounts/' + recipient_account_id
         response = trolley.configuration.Configuration.client(
             self.config).get(endpoint)
-        recipaccount = trolley.recipient_account.RecipientAccount.factory(
+        recipaccount = RecipientAccount.factory(
             response)
         recipientaccount = namedtuple(
             "RecipientAccount", recipaccount.keys())(*recipaccount.values())
         return recipientaccount
 
+    """
+        Create a recipient account
+            A recipient_id and body are required::
+            recipient_account.create('R-4625iLug2GKqKZG2WzAf3e',{"accountHolderName": "Acer" ...})
+        """
     def create(self, recipient_id, body):
         if recipient_id is None:
             raise InvalidFieldException("Recipient id cannot be None.")
         endpoint = '/v1/recipients/' + recipient_id + '/accounts'
         response = trolley.configuration.Configuration.client(
             self.config).post(endpoint, body)
-        recipaccount = trolley.recipient_account.RecipientAccount.factory(
+        recipaccount = RecipientAccount.factory(
             response)
         recipientaccount = namedtuple(
             "RecipientAccount", recipaccount.keys())(*recipaccount.values())
         return recipientaccount
 
+    """
+        Update a recipient account
+            A recipient_id, recipient_account_id, and body are required::
+            recipient_account.update('R-fjeracjmuflh','A-2DQMpN4jurTFn9gRxobx4C',
+            {"accountHolderName": "Acer Philips"})
+        """
     def update(self, recipient_id, recipient_account_id, body):
         if recipient_id is None:
             raise InvalidFieldException("Recipient id cannot be None.")
         endpoint = '/v1/recipients/' + recipient_id + '/accounts/' + recipient_account_id
         response = trolley.configuration.Configuration.client(
             self.config).patch(endpoint, body)
-        recipaccount = trolley.recipient_account.RecipientAccount.factory(
+        recipaccount = RecipientAccount.factory(
             response)
         recipientaccount = namedtuple(
             "RecipientAccount", recipaccount.keys())(*recipaccount.values())
         return recipientaccount
 
+    """
+        Delete a recipient account
+            A recipient_id and recipient_account_id are required::
+            recipient_account.delete('R-fjeracjmuflh','A-2DQMpN4jurTFn9gRxobx4C')
+        """
     def delete(self, recipient_id, recipient_account_id):
         if recipient_id is None:
             raise InvalidFieldException("Recipient id cannot be None.")
